backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
httpx>=0.28.1,<1.0.0
redis==5.0.4
apscheduler==3.10.4
cloudinary==1.40.0
google-genai==1.69.0
Pillow==10.3.0
python-dotenv==1.0.1
slowapi==0.1.9
loguru==0.7.2
sentry-sdk[fastapi]==2.8.0
pytest==8.1.1
pytest-asyncio==0.23.6
aiosqlite==0.20.0
greenlet>=3.0.0,<4.0.0
pgvector==0.3.6
arq==0.26.3
aiosmtplib==3.0.2
jinja2==3.1.4
langchain==0.3.28
langchain-openai==0.3.35
langchain-core==0.3.83
pyotp>=2.9.0,<3.0.0
stripe>=8.0.0,<9.0.0
